requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bitarray==2.7.2
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.12.1
Django==4.1.7
django-redis==5.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
importlib-resources==5.12.0
ipfshttpclient==0.8.0a2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
lru-dict==1.1.8
MarkupSafe==2.1.2
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
openapi-codec==1.3.2
parsimonious==0.8.1
pkgutil_resolve_name==1.3.10
protobuf==3.19.5
psycopg2-binary==2.9.5
pycryptodome==3.17
pyrsistent==0.19.3
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
redis==4.5.1
requests==2.28.2
rlp==2.0.1
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
toolz==0.12.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
varint==1.0.2
web3==5.31.3
websockets==9.1
yarl==1.8.2
zipp==3.14.0
